requirements.txt
psycopg2==2.8.4
python-dateutil==2.8.1
PyVCF==0.6.8
